added speed and changed talon to sparkmax BECAUSE IHATE YOU JAMES I HATE YOU I HATE YOU LEMON WE'RE USING SPARKMAX

+++ src/main/java/frc/robot/Constants.java
@@ -5,7 +5,7 @@ public class Constants {
     public static final int leftKohaiCanID = 1;
     public static final int rightSenpaiCanID = 2;
     public static final int rightKohaiCanID = 3;
-    public static int speed = 1
+    public static double speed = 1
 
     public enum BloodType = {
         A,
